+++ src/ScanPatron/ScanPatron.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ScanPatron = () => (
+  <div>ScanPatron</div>
+);
+
+export default ScanPatron;