+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.0 (IN PROGRESS)
 
+* Add Reading Room Access App into App List. Refs UIRR-3.
 * Update version in package.json
 * Add initial files.
 * New app created with stripes-cli